combine ComputeControllerAPIImplementation and UserComputeAPIImplementation
combined ComputeControllerAPIImplementation and UserComputeAPIImplementation

added the input validation to the methods
